Fix usb_control_msg() returning -EAGAIN.
It was caused by using a stack array as the data buffer, which Linux does not
allow because it is not DMA capable.  Also fix the logic in the UTF-16-to-ASCII
conversion loop.

Also add more helpful error messages to help with diagnosing potential errors,
and fix a few errors in code style.