Avoid sscanf() %s bug & add informative error messages
Setting the led_logo attribute is complicated enough to merit more descriptive
error messages to the user when the input is invalid.

Use better parameters and return values for the leds_store_* functions for ease
of use and helpful error messages.

Also fix not returning the count on success in led_logo_store().