Split up batches into partitions for dynamic LED attributes
The list of batches is split up into partitions of 2, and each batch in a
partition is set by one color (or 8 colors if its the ring LEDs) in the buffer,
at preset "fixed".  This means the buffer -- after the initial source
specification -- only contains a list of colors or color 8-tuples.

This avoids the buffer size exceeding PAGE_SIZE, which causes everything after
PAGE_SIZE bytes in it to be cut off, and thereby lead to a malformed buffer by
no fault of the user.

This of course also results in a lessened degree of control of colors and no
control of other message attributes; everything is set to default with preset
"fixed".