Implement dynamic fan/pump speed update
The speed is set based on a dynamic value and the speed type used.  Use the same
dynamic value source setup as for LED attributes; factor it out into its own
file.  The speed types are the same as in the NXZT Cam program, except type
"custom" gives full control to the user.