Replace error messages by warnings for attr parse errs
Attributes being set to invalid values is not an unanticipated error, but rather
a well-handled use-case, so we shouldn't fill up the logs with garbage error
messages if the user-space API is used as expected.  However, useful warnings
are still present for users that want to figure out how exactly their input is
invalid for the more complicated attributes.

Add a useful info message when restarting updates, to match the info message
when halting updates.