docs(README): explain how to permanently install the drivers
Both using `insmod` and copying the module file to `/lib/modules/...` is now
explained

Also explain a better solution for preventing the issue of `usbhid` connecting
to the device, which is permanent.

Remove the superfluous troubleshooting section concerning the general case of
another module holding the device, and instead just explain how to deal with the
specific case of `usbhid`.  The general case explanation may be re-added later
if a user has such an issue.